Tidied up code and moved phone number presenter into own class and spec
