+++ app/models/crime_application.rb
@@ -1,28 +1,11 @@
 require 'laa_crime_schemas'
 
-# rubocop:disable Metrics/ClassLength
 class CrimeApplication < LaaCrimeSchemas::Structs::CrimeApplication
   include Assignable
   include Reviewable
 
   def supporting_evidence
     super.map { |document| Document.new(document.attributes) }
-  end
-
-  def applicant_name
-    applicant.full_name
-  end
-
-  def formatted_applicant_nino
-    return if applicant.nino.nil?
-
-    applicant.nino
-  end
-
-  def formatted_applicant_telephone_number
-    return if applicant.telephone_number.nil?
-
-    format_telephone_number(applicant.telephone_number)
   end
 
   def legal_rep_name
@@ -113,6 +96,16 @@ class CrimeApplication < LaaCrimeSchemas::Structs::CrimeApplication
     crime_application.id
   end
 
+  def applicant_name
+    applicant.full_name
+  end
+
+  def formatted_applicant_nino
+    return if applicant.nino.nil?
+
+    applicant.nino
+  end
+
   def all_histories
     return @all_histories if @all_histories
 
@@ -126,32 +119,13 @@ class CrimeApplication < LaaCrimeSchemas::Structs::CrimeApplication
     @case_details ||= CaseDetailsPresenter.present(self[:case_details])
   end
 
+  def phone_number
+    @phone_number ||= PhoneNumberPresenter.present(self)
+  end
+
   private
 
   def applicant
     client_details.applicant
   end
-
-  # rubocop:disable Metrics/MethodLength, Metrics/CyclomaticComplexity
-  def format_telephone_number(number)
-    # Remove all spaces
-    formatted_tel = number.gsub(/\s+/, '')
-
-    case formatted_tel
-    when /\A07\d{9}\z/ # Mobile number with exactly 11 digits
-      [5, 9].each { |i| formatted_tel.insert i, ' ' }
-    when /\A\+447\d{9}\z/ # +44 mobile number format without initial 0
-      [3, 8, 12].each { |i| formatted_tel.insert i, ' ' }
-    when /\A\+4407\d{9}\z/ # +44 mobile number format with initial 0
-      [3, 9, 13].each { |i| formatted_tel.insert i, ' ' }
-    when /\A\+44\(0\)7\d{9}\z/ # +44 mobile numberformat with initial 0 in parentheses
-      [3, 7, 12, 16].each { |i| formatted_tel.insert i, ' ' }
-    else
-      formatted_tel = applicant.telephone_number
-    end
-
-    formatted_tel
-  end
-  # rubocop:enable Metrics/MethodLength, Metrics/CyclomaticComplexity
 end
-# rubocop:enable Metrics/ClassLength
