+++ app/views/casework/crime_applications/show.html.erb
@@ -9,7 +9,7 @@
                locals: {
                  applicant: @crime_application.client_details.applicant,
                  nino: @crime_application.formatted_applicant_nino,
-                 applicant_tel: @crime_application.formatted_applicant_telephone_number
+                 applicant_tel: @crime_application.phone_number.for_applicant
                } %>
     <%= render partial: 'case_details', object: @crime_application.case_details %>
     <%= render partial: 'offences', object: @crime_application.case_details.offences %>
