+++ spec/models/crime_application_spec.rb
@@ -25,36 +25,6 @@ RSpec.describe CrimeApplication do
     it { is_expected.to eq 'AJ123456C' }
   end
 
-  describe '#formatted_applicant_telephone_number' do
-    subject(:formatted_applicant_telephone_number) { application.formatted_applicant_telephone_number }
-
-    it { is_expected.to eq '07771 231 231' }
-
-    context 'with input that has +44 and no 0' do
-      before do
-        attributes['client_details']['applicant']['telephone_number'] = '+447123123123'
-      end
-
-      it { is_expected.to eq '+44 7123 123 123' }
-    end
-
-    context 'with input that has +44 and 0' do
-      before do
-        attributes['client_details']['applicant']['telephone_number'] = '+4407123456456'
-      end
-
-      it { is_expected.to eq '+44 07123 456 456' }
-    end
-
-    context 'with input that has +44 and 0 with parentheses' do
-      before do
-        attributes['client_details']['applicant']['telephone_number'] = '+44(0)7123789789'
-      end
-
-      it { is_expected.to eq '+44 (0) 7123 789 789' }
-    end
-  end
-
   describe '#assignee_name' do
     subject(:assignee_name) { application.assignee_name }
